Refactor workspace creation logic and add personal workspace creation on email signup
- Moved workspace creation functions to a new file for better organization.
- Implemented  to create a personal workspace after user signup.
- Removed unused imports and functions from .
- Updated  to call the new personal workspace creation function during email signup.
