+++ apps/sim/lib/auth/auth.ts
@@ -99,6 +99,7 @@ import {
 import { getCanonicalScopesForProvider } from '@/lib/oauth/utils'
 import { captureServerEvent, getPostHogClient } from '@/lib/posthog/server'
 import { disableUserResources } from '@/lib/workflows/lifecycle'
+import { ensurePersonalWorkspaceOnEmailSignup } from '@/lib/workspaces/create-workspace'
 import { SSO_TRUSTED_PROVIDERS } from '@/ee/sso/constants'
 import { createAnonymousSession, ensureAnonymousUserExists } from './anonymous'
 import { getRequestedSignInProviderId, isSignInProviderAllowed } from './constants'
@@ -995,6 +996,40 @@ export const auth = betterAuth({
       }
 
       return
+    }),
+    after: createAuthMiddleware(async (ctx) => {
+      if (!ctx.path.startsWith('/sign-up/email')) {
+        return
+      }
+
+      const newSession = ctx.context.newSession
+      if (!newSession?.user?.id) {
+        return
+      }
+
+      try {
+        const body = ctx.body as Record<string, unknown> | undefined
+        const organizationIdRaw =
+          body?.organizationId ??
+          body?.organization_id ??
+          ctx.headers?.get('x-organization-id') ??
+          ctx.headers?.get('X-Organization-Id')
+        const organizationId =
+          typeof organizationIdRaw === 'string' && organizationIdRaw.trim().length > 0
+            ? organizationIdRaw.trim()
+            : null
+
+        await ensurePersonalWorkspaceOnEmailSignup({
+          userId: newSession.user.id,
+          userName: newSession.user.name,
+          organizationId,
+        })
+      } catch (error) {
+        logger.error('[hooks.after] Failed to create personal workspace after email signup', {
+          userId: newSession.user.id,
+          error,
+        })
+      }
     }),
   },
   plugins: [
